+++ includes/class-wdevs-tax-switch-block-label.php
@@ -39,8 +39,9 @@ class Wdevs_Tax_Switch_Block_Label extends Wdevs_Tax_Switch_Block {
 	 * @since 1.5.0
 	 */
 	public function block_render_callback( $attributes = [], $content = '' ) {
-		if ( ! is_admin() ) {
-			$this->enqueue_frontend_scripts();
+		$shouldRender = $this->before_component_render();
+		if ( ! $shouldRender ) {
+			return '';
 		}
 
 		return $this->add_attributes_to_block( $attributes, $content );
@@ -69,8 +70,9 @@ class Wdevs_Tax_Switch_Block_Label extends Wdevs_Tax_Switch_Block {
 	 * @since 1.5.0
 	 */
 	public function shortcode_render_callback( $attributes = [], $content = '' ) {
-		if ( ! is_admin() ) {
-			$this->enqueue_frontend_scripts();
+		$shouldRender = $this->before_component_render();
+		if ( ! $shouldRender ) {
+			return '';
 		}
 
 		$attributes = shortcode_atts( [